Move middleware logger into a separate file logger.js
