+++ 02-express-tutorial/app.js
@@ -89,21 +89,32 @@ const app = express();
 
 // --------------------------------------
 
-const logger = (req, res, next) => {
-  const method = req.method;
-  const url = req.url;
-  const today = new Date();
-  const date =
-    today.getFullYear() + "-" + (today.getMonth() + 1) + "-" + today.getDate();
-  const time =
-    today.getHours() + ":" + today.getMinutes() + ":" + today.getSeconds();
-  const dateTime = date + " " + time;
+// middleware week 4
 
-  console.log(method, url, dateTime);
-  next();
-};
+// moved to logger.js 
 
-app.use(["/", "/about"], logger);
+const logger = require("./logger");
+// const logger = (req, res, next) => {
+//   const method = req.method;
+//   const url = req.url;
+//   const today = new Date();
+//   const date =
+//     today.getFullYear() + "-" + (today.getMonth() + 1) + "-" + today.getDate();
+//   const time =
+//     today.getHours() + ":" + today.getMinutes() + ":" + today.getSeconds();
+//   const dateTime = date + " " + time;
+
+//   console.log(method, url, dateTime);
+//   next();
+// };
+
+// app.use(["/", "/about"], logger);
+
+// app.get('/',logger, (req, res) => {
+//   res.send('Home')
+// })
+
+app.use(logger); //apply to all urls
 
 app.get('/',(req, res) => {
   res.send('Home')
